fix(sobre-mi): unify profile card and decision cycle in About
Single #sobre-mi block: photo + ficha + cycle/fields/equation.
Remove standalone PerfilEstrategico section (no double headers/photo).
Keep #perfil-estrategico anchor on the cycle strip.
