+++ src/pages/SobreMi.tsx
@@ -1,5 +1,4 @@
 import { About } from '../components/organisms/About';
-import { PerfilEstrategico } from '../components/organisms/PerfilEstrategico';
 import { Skills } from '../components/organisms/Skills';
 import { ProfileScope } from '../components/organisms/ProfileScope';
 import { Experience } from '../components/organisms/Experience';
@@ -43,17 +42,13 @@ const SobreMi = () => {
         url={canonicalFromPath('/sobre-mi')}
       />
       {/*
-        L1 Perfil reclutador (ficha)
-        L1b Perfil estratégico (ciclo de campo — réplica en código, no imagen pegada)
+        L1 Perfil unificado (ficha + ciclo de decisión en un bloque)
         L2 Método + alcance
         L3 Timeline (VN / micro1 con CTA a evidencia)
-        L4 Evidencias ancladas al cargo (no galería general)
-           - #evidencia-vn  → UX Manager VN (Monitas, Edu21, funnels, FO)
-           - #evidencia-micro1 → Anotación + QA grabación
+        L4 Evidencias ancladas al cargo
         Enterprise employment SURA/Transvip → /proyectos
       */}
       <About />
-      <PerfilEstrategico />
       <Skills />
       <ProfileScope />
       <Experience />
